Yadam newrepo (#48)
* workaround configuration issue

* Add changeCursorState function

* Modify TopNav

* Add Alert to CameraConfigureinfo

* Moved defaultCamera & defaultLocation under public/icons

* Fix Play.png Path

* Remove replacement of 8000 to 3000

* django to serve front-end as well

* add dockerfile

* update readme for containerization

* add ui_production

+++ VisionOnEdge/ui/src/components/CameraConfigure/CameraConfigureInfo.tsx
@@ -16,6 +16,9 @@ export const CameraConfigureInfo: React.FC = () => {
   const history = useHistory();
 
   const onDeleteConfigure = (): void => {
+    // eslint-disable-next-line no-restricted-globals
+    const sureDelete = confirm('Delete this configuration?');
+    if (!sureDelete) return;
     const result = (dispatch(thunkDeleteProject(projectId)) as unknown) as Promise<any>;
     result
       .then((data) => {
@@ -42,7 +45,7 @@ export const CameraConfigureInfo: React.FC = () => {
   return (
     <Flex column gap="gap.large">
       <h1>Configuration</h1>
-      {project.modelUrl ? (
+      {!project.modelUrl ? (
         <Loader size="largest" label="Trainning" labelPosition="below" design={{ paddingTop: '300px' }} />
       ) : (
         <>
